Changed migrations, changed some methods

+++ EPlast/EPlast.WebApi/Controllers/KadraVykhovnykivController.cs
@@ -1,7 +1,6 @@
 ﻿using EPlast.BLL.DTO.EducatorsStaff;
 using EPlast.BLL.Interfaces.EducatorsStaff;
 using EPlast.BLL.Interfaces.Logging;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System;
@@ -11,8 +10,6 @@ namespace EPlast.WebApi.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    [Authorize(AuthenticationSchemes = "Bearer")]
-    [Authorize(Roles = "Admin")]
     public class KadraVykhovnykivController : ControllerBase
     {
         private readonly ILoggerService<KadraVykhovnykivController> _logger;
@@ -35,9 +32,18 @@ namespace EPlast.WebApi.Controllers
         [HttpPost("CreateKadra")]
         public async Task<IActionResult> CreateKadra(KadraVykhovnykivDTO kvDTO)
         {
+                if (User.IsInRole("Admin"))
+                {
                     var newKadra=await _kvService.CreateKadra(kvDTO);
-                   
+                    _logger.LogInformation($"User {{{kvDTO.UserId}}} gained Kadra Vykhovnykiv of type: {{{kvDTO.KVTypesID}}}");
                 return Ok(newKadra);
+                }
+                else
+                {
+                _logger.LogError("Current user is not an admin");
+                return StatusCode(StatusCodes.Status403Forbidden);
+                }
+              
         }
 
         /// <summary>
@@ -47,12 +53,27 @@ namespace EPlast.WebApi.Controllers
         /// <response code="200">Successful operation</response>
         /// <response code="403">User is not admin</response>
         ///  <response code="404">kadra with this id doesn't exist</response>
-        [HttpDelete("RemoveKadra/{kadraId}")]
+        [HttpDelete("RemoveKadra/{kadra_id:int}")]
         public async Task<IActionResult> Remove(int kadraId)
         {
-
+            try
+            {
+                if (User.IsInRole("Admin"))
+                {
                     await _kvService.DeleteKadra(kadraId);
                     return StatusCode(StatusCodes.Status200OK);
+                }
+                else
+                {
+                    _logger.LogError("Current user is not an admin");
+                    return StatusCode(StatusCodes.Status403Forbidden);
+                }
+            }
+            catch( InvalidOperationException e)
+            {
+                _logger.LogError(e.Message);
+                return StatusCode(StatusCodes.Status404NotFound);
+            }
            
         }
 
@@ -66,9 +87,17 @@ namespace EPlast.WebApi.Controllers
         [HttpPut("EditKadra")]
         public async Task<IActionResult> Update( KadraVykhovnykivDTO kadrasDTO)
         {
-
+           
+            if (User.IsInRole("Admin"))
+            {
                 await _kvService.UpdateKadra(kadrasDTO);
                 return StatusCode(StatusCodes.Status200OK);
+            }
+            else
+            {
+                _logger.LogError("Current user is not an admin");
+                return StatusCode(StatusCodes.Status403Forbidden);
+            } 
         }
 
 
@@ -158,6 +187,8 @@ namespace EPlast.WebApi.Controllers
             
             
         }
+
+
 
 
     }
